+++ src/GafferSceneModule/GlobalsBinding.cpp
@@ -39,7 +39,7 @@
 #include "GlobalsBinding.h"
 
 #include "GafferScene/DeleteOutputs.h"
-#include "GafferScene/DeletePasses.h"
+#include "GafferScene/DeleteRenderPasses.h"
 #include "GafferScene/DeleteSets.h"
 #include "GafferScene/GlobalShader.h"
 #include "GafferScene/Outputs.h"
@@ -116,11 +116,11 @@ void GafferSceneModule::bindGlobals()
 	DependencyNodeClass<GlobalShader>();
 
 	{
-		scope s = DependencyNodeClass<DeletePasses>();
+		scope s = DependencyNodeClass<DeleteRenderPasses>();
 
-		enum_<DeletePasses::Mode>( "Mode" )
-			.value( "Keep", DeletePasses::Keep )
-			.value( "Delete", DeletePasses::Delete )
+		enum_<DeleteRenderPasses::Mode>( "Mode" )
+			.value( "Keep", DeleteRenderPasses::Keep )
+			.value( "Delete", DeleteRenderPasses::Delete )
 		;
 	}
 	DependencyNodeClass<RenderPasses>();